+++ activerecord/CHANGELOG.md
@@ -1,3 +1,20 @@
+*   Support merging option `:rewhere` to allow mergee side condition to be replaced exactly.
+
+    ```ruby
+    david_and_mary = Author.where(id: david.id..mary.id)
+
+    # both conflict conditions exists
+    david_and_mary.merge(Author.where(id: bob)) # => []
+
+    # mergee side condition is replaced by rewhere
+    david_and_mary.merge(Author.rewhere(id: bob)) # => [bob]
+
+    # mergee side condition is replaced by rewhere option
+    david_and_mary.merge(Author.where(id: bob), rewhere: true) # => [bob]
+    ```
+
+    *Ryuta Kamizono*
+
 *   Add support for finding records based on signed ids, which are tamper-proof, verified ids that can be
     set to expire and scoped with a purpose. This is particularly useful for things like password reset
     or email verification, where you want the bearer of the signed id to be able to interact with the